move pem module to own file
